move unique hash key to Request model, polish code
